requirements.txt
flask>=2.0.1
python-dotenv==0.19.0
openai==1.3.0
google-generativeai==0.3.1
numpy>=1.24.0
scikit-learn>=1.0.2
requests==2.31.0
werkzeug<3.0.0
